requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
cfgv==3.4.0
click==8.2.1
coverage==7.8.2
distlib==0.3.9
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
filelock==3.18.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
identify==2.6.12
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
nodeenv==1.9.1
packaging==25.0
platformdirs==4.3.8
pluggy==1.6.0
pre_commit==4.2.0
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
pytest==8.4.0
pytest-check==2.5.3
pytest-cov==6.1.1
pytest-env==1.1.5
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
rich==14.0.0
rich-toolkit==0.14.7
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.34.3
uvloop==0.21.0
virtualenv==20.31.2
watchfiles==1.0.5
websockets==15.0.1
